Удален private MessageHandling bot
